Customize OJS config and add placeholders
This ports the setup from the previous repo, but with all the settings
from OJS (not sure if they were deleted by accident or what), and adds
%FOO% style placeholders we can adjust via an in-container script so
each environment can have its own settings.

The rename to `.ini` is just so it's easier to edit. Not sure why they
didn't just go with an INI file to begin with....