docker: don't lose the config on restarts!
Fixes the issue where `config.inc.php` is gone, and then replaced, after
a "stop" / "down" type of compose command. OJS was replacing the file
with the default which caused the install UI to show up repeatedly.

The config file only needs to have replacements done once (for now), but
it must be re-linked into the app dir every startup since the symlink is
just an in-container file.