Refactor the whole compose setup
This switches our approach to rely on includes rather than a single core
compose where we have to specify things like the entire DB service in an
override. We should be more able to set up and combine services as
needed, especially when working on a migration image and container.