Provide compose example for multiple allowed hosts
The extra quoting fixes the unexplained 500 error when trying to load
the site.

The error basically says something about an array being expected, which
turns out to be because of the allowed hosts value being treated as a
string unless the actual environment value is put inside single quotes.

This is due to a problem because we use an environment setting which
then gets stuffed into `config.inc.php`. In yaml, defining `FOO: 'bar'`
results in FOO being just the three characters, `bar`, not the
surrounding single-quotes. If those quotes are a required part of the
value, you have to actually quote them, e.g., `FOO: "'bar'"`.

For some reason, OJS *must* have its array values inside literal quotes.
Except that it works fine when there's only one element. Having hosts
set to `["localhost"]` is fine. `["localhost", "127.0.0.1"]` is not.
`'["localhost, "127.0.0.1"]'` is the fix.

My brain hurts.